Time: 20 ms (100%), Space: 58.2 MB (21.88%) - LeetHub

+++ 1335-maximum-candies-allocated-to-k-children/1335-maximum-candies-allocated-to-k-children.cs
@@ -24,7 +24,7 @@ public class Solution {
         }
         
         }
-       return (int)right;
+       return right;
         
     }
 }
